setup controllers/routes and break out consent for consistency
